+++ jagua-rs/src/geometry/original_shape.rs
@@ -54,7 +54,7 @@ impl OriginalShape {
         self.shape.centroid()
     }
 
-    pub fn area(&self) -> f32 {
+    pub fn area(&self) -> f64 {
         self.shape.area
     }
 
@@ -62,7 +62,7 @@ impl OriginalShape {
         self.shape.bbox
     }
 
-    pub fn diameter(&self) -> f32 {
+    pub fn diameter(&self) -> f64 {
         self.shape.diameter
     }
 }
