+++ jagua-rs/src/probs/qpp/entities/instance.rs
@@ -22,10 +22,10 @@ impl QPInstance {
         Self { items, base_square }
     }
 
-    pub fn item_area(&self) -> f32 {
+    pub fn item_area(&self) -> f64 {
         self.items
             .iter()
-            .map(|(item, qty)| item.shape_orig.area() * *qty as f32)
+            .map(|(item, qty)| item.shape_orig.area() * *qty as f64)
             .sum()
     }
