+++ jagua-rs/src/probs/qpp/io/ext_repr.rs
@@ -25,11 +25,11 @@ pub struct ExtItem {
 #[derive(Serialize, Deserialize, Clone)]
 pub struct ExtQPSolution {
     /// The strip width of the solution
-    pub square_side: f32,
+    pub square_side: f64,
     /// Layouts which compose the solution
     pub layout: ExtLayout,
     /// Sum of the area of the produced items divided by the sum of the area of the containers
-    pub density: f32,
+    pub density: f64,
     /// The time it took to generate the solution in seconds
     pub run_time_sec: u64,
 }
