+++ lbf/src/config.rs
@@ -10,19 +10,19 @@ pub struct LBFConfig {
     /// Configuration of the Collision Detection Engine
     pub cde_config: CDEConfig,
     /// Max deviation from the original polygon area as a fraction. If undefined, the algorithm will run without simplification
-    pub poly_simpl_tolerance: Option<f32>,
+    pub poly_simpl_tolerance: Option<f64>,
     /// Maximum distance between two vertices of a polygon to consider it a narrow concavity (which will be closed).
     /// Defined as a fraction of the largest item in the instance.
-    pub narrow_concavity_cutoff_ratio: Option<f32>,
+    pub narrow_concavity_cutoff_ratio: Option<f64>,
     /// Minimum distance between items and other hazards.
     /// If undefined, the algorithm will run without this constraint
-    pub min_item_separation: Option<f32>,
+    pub min_item_separation: Option<f64>,
     /// Seed for the PRNG. If undefined, the algorithm will run in non-deterministic mode using entropy
     pub prng_seed: Option<u64>,
     /// Total budget of samples per item per layout
     pub n_samples: usize,
     /// Fraction of `n_samples_per_item` used for the local search sampler, the rest is sampled uniformly.
-    pub ls_frac: f32,
+    pub ls_frac: f64,
     /// Optional SVG drawing options
     pub svg_draw_options: SvgDrawOptions,
 }
