+++ lbf/src/opt/loss.rs
@@ -2,7 +2,7 @@ use std::cmp::Ordering;
 
 use jagua_rs::geometry::primitives::{Rect, SPolygon};
 
-const X_MULTIPLIER: f32 = 10.0;
+const X_MULTIPLIER: f64 = 10.0;
 
 /// The loss LBF assigned to a placing option.
 /// Weighted sum of the x_max and y_max of the shape, with the horizontal dimension being more important.
@@ -10,8 +10,8 @@ const X_MULTIPLIER: f32 = 10.0;
 /// A pure lexicographic comparison (always prioritizing x-axis) would lead to undesirable results due to the continuous nature of the values.
 #[derive(Copy, Clone, Debug, PartialEq)]
 pub struct LBFLoss {
-    x_max: f32,
-    y_max: f32,
+    x_max: f64,
+    y_max: f64,
 }
 
 impl LBFLoss {
@@ -26,7 +26,7 @@ impl LBFLoss {
         LBFLoss::from_bbox(shape.bbox)
     }
 
-    pub fn cost(&self) -> f32 {
+    pub fn cost(&self) -> f64 {
         self.x_max * X_MULTIPLIER + self.y_max
     }
 
@@ -36,7 +36,7 @@ impl LBFLoss {
         let x_max_bound = cost / X_MULTIPLIER;
 
         let mut tightened_bbox = sample_bbox;
-        tightened_bbox.x_max = f32::min(sample_bbox.x_max, x_max_bound);
+        tightened_bbox.x_max = f64::min(sample_bbox.x_max, x_max_bound);
 
         tightened_bbox
     }
