+++ lbf/src/samplers/ls_sampler.rs
@@ -1,7 +1,7 @@
 use rand::Rng;
 use rand_distr::Distribution;
 use rand_distr::Normal;
-use std::f32::consts::PI;
+use std::f64::consts::PI;
 
 use crate::samplers::rotation_distr::NormalRotDistr;
 use jagua_rs::entities::Item;
@@ -9,22 +9,22 @@ use jagua_rs::geometry::DTransformation;
 use jagua_rs::geometry::primitives::Rect;
 
 /// The stddev of translation starts at 1% and ends at 0.05% of the largest dimension of the bounding box.
-pub const SD_TRANSL: (f32, f32) = (0.01, 0.0005);
+pub const SD_TRANSL: (f64, f64) = (0.01, 0.0005);
 
 /// The stddev of rotation starts at 2° and ends at 0.5°.
-pub const SD_ROT: (f32, f32) = (2.0 * PI / 180.0, 0.5 * PI / 180.0);
+pub const SD_ROT: (f64, f64) = (2.0 * PI / 180.0, 0.5 * PI / 180.0);
 
 ///Creates `Transformation` samples for a given item.
 ///The samples are drawn from normal distributions with decaying standard deviations.
 ///Each time an improvement is found, the mean of the distributions is shifted to the new best transformation.
 pub struct LSSampler {
-    normal_x: Normal<f32>,
-    normal_y: Normal<f32>,
+    normal_x: Normal<f64>,
+    normal_y: Normal<f64>,
     normal_r: NormalRotDistr,
-    sd_transl: f32,
-    sd_rot: f32,
-    sd_transl_range: (f32, f32),
-    sd_rot_range: (f32, f32),
+    sd_transl: f64,
+    sd_rot: f64,
+    sd_transl_range: (f64, f64),
+    sd_rot_range: (f64, f64),
     pub(crate) n_samples: usize,
 }
 
@@ -32,8 +32,8 @@ impl LSSampler {
     pub fn new(
         item: &Item,
         ref_transform: DTransformation,
-        sd_transl_range: (f32, f32),
-        sd_rot_range: (f32, f32),
+        sd_transl_range: (f64, f64),
+        sd_rot_range: (f64, f64),
     ) -> Self {
         let sd_transl = sd_transl_range.0;
         let sd_rot = sd_rot_range.0;
@@ -56,7 +56,7 @@ impl LSSampler {
 
     /// Creates a new sampler with default standard deviation ranges: [SD_TRANSL] and [SD_ROT].
     pub fn from_defaults(item: &Item, ref_transform: DTransformation, bbox: Rect) -> Self {
-        let max_dim = f32::max(bbox.width(), bbox.height());
+        let max_dim = f64::max(bbox.width(), bbox.height());
         let sd_transl_range = (SD_TRANSL.0 * max_dim, SD_TRANSL.1 * max_dim);
         Self::new(item, ref_transform, sd_transl_range, SD_ROT)
     }
@@ -69,7 +69,7 @@ impl LSSampler {
     }
 
     /// Sets the standard deviation of the normal distributions.
-    pub fn set_stddev(&mut self, stddev_transl: f32, stddev_rot: f32) {
+    pub fn set_stddev(&mut self, stddev_transl: f64, stddev_rot: f64) {
         assert!(stddev_transl >= 0.0 && stddev_rot >= 0.0);
 
         self.sd_transl = stddev_transl;
@@ -86,8 +86,8 @@ impl LSSampler {
     /// f(0) = init;
     /// f(1) = end;
     /// f(x) = init * (end/init)^x;
-    pub fn decay_stddev(&mut self, progress_pct: f32) {
-        let calc_stddev = |(init, end): (f32, f32), pct: f32| init * (end / init).powf(pct);
+    pub fn decay_stddev(&mut self, progress_pct: f64) {
+        let calc_stddev = |(init, end): (f64, f64), pct: f64| init * (end / init).powf(pct);
         self.set_stddev(
             calc_stddev(self.sd_transl_range, progress_pct),
             calc_stddev(self.sd_rot_range, progress_pct),
