+++ lbf/src/util/assertions.rs
@@ -5,8 +5,8 @@ pub fn strip_width_is_in_check(prob: &SPProblem) -> bool {
         .instance
         .items
         .iter()
-        .map(|(i, q)| i.shape_cd.diameter * *q as f32)
-        .sum::<f32>();
+        .map(|(i, q)| i.shape_cd.diameter * *q as f64)
+        .sum::<f64>();
 
     prob.strip_width() < 2.0 * (diameters_of_all_items)
 }
